requirements.txt
fastapi
uvicorn
tortoise-orm[asyncpg]
pydantic
python-dotenv
openai
requests